+++ ts/packages/aiclient/src/runtimeConfig.ts
@@ -22,27 +22,20 @@ import { setActiveModelProvider } from "./providerMode.js";
 
 let cached: Config | undefined;
 
-// When the active provider is Copilot, eagerly start the CLI client so the
-// first user-visible request doesn't pay the one-time spawn/startup cost.
-// Dynamic import avoids a static import cycle (copilotModels -> copilotSettings
-// -> runtimeConfig); fire-and-forget so host startup is never blocked.
-function maybePrewarmCopilot(config: Config): void {
-    if (config.modelProvider !== "copilot") return;
-    void import("./copilotModels.js")
-        .then((m) => m.warmupCopilotFromConfig())
-        .catch(() => {});
-}
-
 /**
  * Replace the cached typed Config. Hosts should call this after they
  * finish populating `process.env` and before any aiclient operation.
+ *
+ * Copilot pre-warm is intentionally NOT triggered here: this low-level
+ * config module must not reach up into the model layer (`copilotModels`),
+ * or it forms an import cycle. Hosts call `warmupCopilotFromConfig()`
+ * themselves after installing the config (see the aiclient barrel export).
  */
 export function setRuntimeConfig(config: Config): void {
     cached = config;
     if (config.modelProvider !== undefined) {
         setActiveModelProvider(config.modelProvider);
     }
-    maybePrewarmCopilot(config);
 }
 
 /**
@@ -58,7 +51,6 @@ export function initRuntimeConfigFromProcessEnv(): Config {
     if (cached.modelProvider !== undefined) {
         setActiveModelProvider(cached.modelProvider);
     }
-    maybePrewarmCopilot(cached);
     return cached;
 }
